requirements.txt
Flask>=2.3,<4.0
python-dotenv>=1.0.0
